Substantial cleanup of size/gen functions
Rewrote these from scratch based on a couple of traits rather than shoe-
horning them onto the side of the regex type itself.

I _think_ the new version of gen_at does a much better job avoiding
reallocations and zeroing out memory for intermediate steps. I at least
noticed a lot of blatant leaks in the old version, that are not present
in the new version. But who knows.