Shockingly painful refactor to support word lists
The root of the trouble was that I needed to be able to run against a
`Vec<String>` or a `&[&str]` depending on whether the words came from
the config or not.

This also does better (?) config handling, removing the tilde expansion.
I shopped around a bit for cross-platform config packages but couldn’t
find any I liked; the ones that handle Windows correctly tend to also
stuff things under Application Support on macOS, and I want to version
my password schemas in `~/.config` on both Linux and Mac.

This also unceremoniously changes the name of the config dir to onepass
to maybe avoid some namespace collisions.