Clearer variable names
Rust-analyzer started showing inferred types everywhere in neovim and I
don’t know how to make it stop; this has been making it harder to know
how my code will look without the LSP stuff on.

The current commit attempts to make clearer what is going on with the
three let bindings for the word list.