New salt format, url canonicalization, usernames
Sites are now canonicalized to URLs, with "https" scheme if no scheme is
present. This also involved a rewrite of the config logic; in the end I
wound up not using toml since URLs as keys are awkward in toml.

The new config format is backwards-incompatible with the old one in that
sites are now represented as key-value pairs rather than list items, and
the new version will generate different passwords from prior versions; I
decided to emphasize this by changing the salt format to use a comma
instead of a colon to separate the increment from the URL.

As well, this version will dump the raw salt to the console in verbose
mode to aid in debugging.

While I was at it, I changed the clap path args to Box<Path> in
preparation for loading paths from the config file, which I hope to do
later.

This also updates the README to match the current functionality, and to
try to make it slightly more approachable now that I’m a little more
settled on at least the salt format.