Breaking word list changes
This enforces some very basic sanity checking on word lists:

- Words are now chosen in alphabetical, rather than file, order. (This
  may break passwords generated from word lists that were not already in
  alphabetical order.)
- Duplicate words are now dropped. (Words that are prefixes of other
  words may still compromise entropy estimates.)
- The empty word is also dropped.

The upshot is that the entropy estimates provided by the app should be
accurate in more cases now than before. N.B. though, we currently fail
open — if we see a word list with duplicates, we do not yet alert the
user; we merely generate passwords from the smaller universe of unique
words.