Change the way we do attestations, v1.0.1
Attestations happen all in one place now. I think this is fine, but it
assumes that artifacts for a workflow run can’t be injected by steps
outside of version control.

Does this in a dedicated job because the github release job already had
contents write permission and I did not want to give the attestation
step that permission.