Add provisioning profile
I just discovered that this needs to be installed on the system in order
for the signed onepass binaries to run on that system. Maybe I can
figure out how to get that to happen with the .pkg, short of
distributing a full app bundle. For now, I should at least put it up for
people to install manually.