use environment variables only for dirs
Now, the config base dir is `$XDG_CONFIG_DIR` if that is set, or else
`%APPDATA%` on Windows. If those are not set, then the config base dir
is `.config` in the user’s home directory (`%USERPROFILE%` on Windows if
that is set, or else `$HOME`.)

The tilde expansion for the words file no longer expands other users’
home directories, and for the time being, paths like `~dir/path` are
treated as invalid.

As a result of this change, we no longer directly depend on nix.