Support including config files
Extends the config file format with an `include:` block that takes a
list of strings. Each string is included, with the included config
overriding the base config. This allows e.g. a
`.config/onepass/local.yaml` that can specify private sites while still
allowing the base configuration to be checked into source control.