Introduce Cow in path manipulation
Reduces the number of string copies needed in the path logic in some
cases. In particular `expand_home` can return the original path in cases
where no modification was done. This is incredibly minor, but is an
actual use for `Cow`, which I had not until now been able to understand
how to use at all for onepass.