Support config entries without schemas (#74)
This was a simple oversight in the SiteConfig parsing code: a site
needed to have a schema or it would be considered invalid.

Instead of that, we now expose a `Config::site_schema` that returns the
schema for a `SiteConfig`, or the default schema.

This also allows sites with no configuration specified, e.g. as a list
for informational purposes.

At this point the config code is quite sprawl-y and probably due for a
refactor.